pyproject.toml
[tool.poetry]
name = "discord-buildathon-2023"
version = "0.1.0"
description = ""
authors = ["EggsLeggs <36281195+EggsLeggs@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-cord = "^2.4.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
